requirements.txt
sentinelsat==1.1.1
xarray==2022.12.0
s3fs==2022.11.0
python-dotenv==0.21.0
